added base architecture to followe all around the project
