+++ lib/src/config/README.md
@@ -1,0 +1,5 @@
+In this folder goes everything general config as
+- Colors
+- Styles
+- Locale
+- Utils
