+++ lib/src/data/README.md
@@ -1,0 +1,3 @@
+This is the persistence  layer for:
+- Repositories
+- Services
