+++ lib/src/ui/README.md
@@ -1,0 +1,1 @@
+This is the visual layer for every screen and widget
